+++ src/heatmap-component/demo.ts
@@ -1,4 +1,4 @@
-import * as d3 from './d3-modules';
+import { Color } from './data/color';
 import { Heatmap } from './heatmap';
 import { formatDataItem } from './utils';
 
@@ -30,7 +30,7 @@ export function demo1(divElementOrId: HTMLDivElement | string) {
         y: d => d.row,
         filter: (d, x, y, xIndex, yIndex) => d.score > 0,
     });
-    const colorScale = d3.scaleLinear([0, 0.5, 1], ['#eeeeee', 'gold', 'red']);
+    const colorScale = Color.createScale([0, 0.5, 1], ['#eeeeee', 'gold', 'red']);
     hm.setColor(d => colorScale(d.score));
     hm.setTooltip((d, x, y, xIndex, yIndex) => `<div style="font-weight: bold; margin-bottom: 0.5em;">${formatDataItem(d)}</div>Column ${x}, Row ${y}<br>Indices [${xIndex},${yIndex}]`);
     setTimeout(() => hm.setFilter(undefined), 2000);
@@ -68,10 +68,10 @@ export function demo1(divElementOrId: HTMLDivElement | string) {
             // console.log('zooming nothing');
         } else {
             // console.log('zooming', e.xMinIndex, e.xMaxIndex, 'values', e.xMin, e.xMax, e);
-            d3.select('#xminindex').text(e.xMinIndex);
-            d3.select('#xmaxindex').text(e.xMaxIndex);
-            d3.select('#xmin').text(e.xMin);
-            d3.select('#xmax').text(e.xMax);
+            setTextContent('#xminindex', e.xMinIndex);
+            setTextContent('#xmaxindex', e.xMaxIndex);
+            setTextContent('#xmin', e.xMin);
+            setTextContent('#xmax', e.xMax);
         }
     });
     hm.setZooming({ axis: 'x' });
@@ -79,13 +79,19 @@ export function demo1(divElementOrId: HTMLDivElement | string) {
     (window as any).hm = hm;
 }
 
+/** Set text content to all HTML elements selected by `elementSelector`.
+ * Example: `setTextContent('#element-to-change', 'changed text here');` */
+function setTextContent(elementSelector: string, content: unknown): void {
+    const elements = document.querySelectorAll(elementSelector);
+    elements.forEach(element => element.textContent = `${content}`);
+}
+
 /** Demo showing a big data example (200_000 x 20) */
 export function demo2(divElementOrId: HTMLDivElement | string) {
     const hm = Heatmap.createDummy(2e5, 20); // Heatmap<number, number, number>
     hm.setVisualParams({ xGapRelative: 0, yGapRelative: 0 });
-    // hm.setVisualParams({ yGapRelative: 0 });
     // hm.setColor(Color.createScale([0, 0.5, 1], ['#00d', '#ddd', '#d00']));
-    // hm.setColor(d3.scaleLinear([0, 0.5, 1], ['#00d', '#ddd', '#d00']));
+    hm.setColor(Color.createScale('Magma', [0, 1], [1, 0]));
     hm.render(divElementOrId);
     // hm.setFilter(d => d > 0.1);
     // setTimeout(()=> hm.setFilter(undefined), 2000);
@@ -118,7 +124,7 @@ export function demo3(divElementOrId: HTMLDivElement | string) {
         x: d => d.col,
         y: d => d.row,
     });
-    const colorScale = d3.scaleLinear([-1, 0, 1], ['#E13D3D', '#ffffff', '#2C8C11']);
+    const colorScale = Color.createScale([-1, 0, 1], ['#E13D3D', '#ffffff', '#2C8C11']); // like d3.scaleLinear([-1, 0, 1], ['#E13D3D', '#ffffff', '#2C8C11']);
     hm.setColor(d => colorScale(d.score));
     hm.setVisualParams({ xGapRelative: 0.1, yGapRelative: 0.1, xGapPixels: null, yGapPixels: null });
     hm.render(divElementOrId);