+++ CHANGELOG.md
@@ -4,11 +4,13 @@ All notable changes to this project will be documented in this file, following t
 
 ## [Unreleased]
 
+-   Support for `Heatmap.createEmpty()` and setting data later
+-   Can manipulate markers via `hm.extensions.marker?.drawMarkers({...})`
+
 ## [0.9.0] - 2024-04-29
 
 -   Renamed the package from `hotmap` to `heatmap-component`
 -   Zooming is off by default (call `hm.setZooming({ axis: 'x' })` to turn it on)
--   Can manipulate markers via `hm.extensions.marker?.drawMarkers({...})`
 
 ## [0.8.0] - 2024-04-26